deleted organized files
moved code into single file
