--- styles/index.css
@@ -1,6 +1,0 @@
-@import url('./reset.css');
-@import url('./variables.css');
-@import url('./typography.css');
-@import url('./layout.css');
-@import url('./fcp.css');
-@import url('./global.css');
