--- styles/layout.css
@@ -1,152 +1,0 @@
-html,
-body {
-  height: 100%;
-
-  &:is(body) {
-    display: flex;
-    flex-direction: column;
-
-    & main {
-      flex: 1;
-    }
-  }
-}
-
-.section:not(#nav .section, .footer .section) {
-  width: 100%;
-  display: grid;
-  grid-auto-flow: row;
-  grid-template-columns: var(--section-template-columns);
-
-  & > *:not(.hero-wrapper) {
-    padding-block: var(--spacing-8);
-
-    &:is(:has(.details-columns)) {
-      padding-block-end: 0;
-    }
-  }
-
-  .full-width {
-    grid-column: full-width;
-  }
-
-  .breakout-left {
-    grid-column: breakout-left;
-
-    .default-content-wrapper :first-child:is(h1, h2, h3, h4, h5, h6) {
-      margin-block-start: 0;
-    }
-  }
-
-  .breakout-right {
-    grid-column: breakout-right;
-
-    .default-content-wrapper :first-child:is(h1, h2, h3, h4, h5, h6) {
-      margin-block-start: 0;
-    }
-  }
-
-  .sidebar-left {
-    grid-column: sidebar-left;
-  }
-
-  .sidebar-right {
-    grid-column: sidebar-right;
-  }
-
-  .content {
-    grid-column: content;
-  }
-
-  .content-left {
-    grid-column: content-left;
-  }
-
-  .content-middle {
-    grid-column: content-middle;
-
-    .default-content-wrapper:not(header .default-content-wrapper, footer .default-content-wrapper) {
-      grid-column: content-middle;
-    }
-  }
-
-  .content-right {
-    grid-column: content-right;
-  }
-
-  .subgrid {
-    display: grid;
-    grid-template-columns: subgrid;
-    grid-column: full-width;
-  }
-}
-
-.default-content-wrapper  {
-  grid-column: content;
-
-  &:is(main .default-content-wrapper) {
-    .button-container {
-      display: inline-flex;
-
-      & + .button-container {
-        margin-inline-start: var(--spacing-4);
-      }
-    }
-  }
-
-  &:is(.section.content > .default-content-wrapper),
-  &:is(body.content .default-content-wrapper) {
-    grid-column: content;
-  }
-
-  @media (width > 1024px) {
-    &:is(.section.content-middle > .default-content-wrapper),
-    &:is(body.content-middle .default-content-wrapper) {
-      grid-column: content-middle;
-    }
-  }
-}
-
-body.details {
-  @media (width > 768px) {
-    .default-content-wrapper:not(header .default-content-wrapper, footer .default-content-wrapper) {
-      grid-column: content-middle / content-right;
-      margin-inline-start: var(--spacing-8);
-    }
-  }
-
-  @media (width > 1024px) {
-    .bento-wrapper {
-      margin-block: var(--spacing-4);
-    }
-
-    .default-content-wrapper:not(header .default-content-wrapper, footer .default-content-wrapper) {
-      grid-column: content-middle / content-right;
-      margin-inline-start: var(--spacing-8);
-    }
-  }
-}
-
-.include-wrapper {
-  display: grid;
-  grid-template-columns: subgrid;
-  grid-column: full-width;
-
-  & > .include.cards {
-    display: grid;
-    grid-template-columns: subgrid;
-    grid-column: content;
-  }
-
-  /* this may have caused a weird breaking issue on mobile safari where the width isn't calculated correctly until rotating the device */
-  .cards.wrapper {
-    display: grid;
-    grid-column: content;
-  }
-
-  &:has(.breakout) {
-    display: grid;
-    grid-column: full-width;
-    grid-template-columns: subgrid;
-  }
-}
