--- styles/reset.css
@@ -1,19 +1,0 @@
-* {
-  box-sizing: border-box;
-
-  &:not(img, picture, helix-sidekick, input, textarea, br, svg, rect, circle, header, footer):empty {
-    display: none;
-  }
-}
-
-html, body, ul, ol {
-  margin: 0;
-  padding: 0;
-}
-
-img {
-  max-width: 100%;
-  height: auto;
-  vertical-align: middle;
-  font-style: italic;
-}
