--- styles/typography.css
@@ -1,45 +1,0 @@
-html, body {
-  font-family: var(--font-copy);
-  font-weight: 400;
-  font-size: var(--font-size);
-}
-
-h1, h2, h3, h4, h5, h6 {
-  font-family: var(--font-display);
-  margin-block: var(--spacing-3);
-  color: var(--color, var(--color-text));
-
-  &:is(h1) { font-size: var(--font-xl); }
-  &:is(h2) { font-size: var(--font-lg); }
-  &:is(h3) { font-size: var(--font-md); }
-  &:is(h4) { font-size: var(--font-sm); }
-  &:is(h5) { font-size: var(--font-xs); }
-
-  & strong {
-    font-weight: 700;
-  }
-}
-
-p { 
-  max-width: 90ch;
-  margin-block: 0;
-}
-
-main {
-  & p {
-    margin-block: var(--spacing-4);
-    color: var(--color, var(--color-text));
-  }
-
-  & ul {
-    color: var(--color, var(--color-text));
-  }
-}
-
-.font-xl { font-size: var(--font-xl); }
-.font-lg { font-size: var(--font-lg); }
-.font-md { font-size: var(--font-md); }
-.font-sm { font-size: var(--font-sm); }
-.font-xs { font-size: var(--font-xs); }
-.font-copy { font-family: var(--font-copy); }
-.font-display { font-family: var(--font-display); }
