--- styles/variables.css
@@ -1,114 +1,0 @@
-:root {
-  --spacing-0: 0rem;
-  --spacing-0-125: 0.125rem;
-  --spacing-0-2: 0.2rem;
-  --spacing-1: 0.25rem;
-  --spacing-2: 0.5rem;
-  --spacing-3: 0.75rem;
-  --spacing-4: 1rem;
-  --spacing-5: 1.25rem;
-  --spacing-6: 1.5rem;
-  --spacing-8: 2rem;
-  --spacing-10: 2.5rem;
-  --spacing-12: 3rem;
-  --spacing-16: 4rem;
-  --spacing-20: 5rem;
-  --spacing-24: 6rem;
-  --spacing-32: 8rem;
-  --spacing-40: 10rem;
-  --spacing-48: 12rem;
-  --spacing-56: 14rem;
-  --spacing-64: 16rem;
-  --spacing-80: 20rem;
-  --spacing-96: 24rem;
-  --spacing-120: 30rem;
-  --spacing-140: 35rem;
-  --spacing-160: 40rem;
-  --spacing-180: 45rem;
-  --spacing-192: 48rem;
-  --spacing-256: 64rem;
-  --spacing-320: 80rem;
-  --spacing-360: 90rem;
-  --spacing-400: 100rem;
-  --spacing-480: 120rem;
-  --layout-max-width: var(--spacing-320);
-  --color-brand-1-hsl: 74 32% 32%;
-  --color-brand-2-hsl: 88 38% 15%;
-  --color-brand-3-hsl: 52 94% 94%;
-  --color-brand-3-hover-hsl: 52 94% 100%;
-  --color-brand-4-hsl: 32 65% 62%;
-  --color-brand-5-hsl: 28 67% 39%;
-  --color-brand-1: hsl(var(--color-brand-1-hsl));
-  --color-brand-2: hsl(var(--color-brand-2-hsl));
-  --color-brand-3: hsl(var(--color-brand-3-hsl));
-  --color-brand-3-hover: hsl(var(--color-brand-3-hover-hsl));
-  --color-brand-4: hsl(var(--color-brand-4-hsl));
-  --color-brand-5: hsl(var(--color-brand-5-hsl));
-  --color-brand: var(--color-brand-1);
-  --color-brand-hsl: var(--color-brand-1-hsl);
-  --color-text-on-brand: var(--color-brand-3);
-  --color-background: var(--color-brand-3);
-  --color-background-hsl: var(--color-brand-3-hsl);
-  --color-foreground: var(--color-brand-4);
-  --color-foreground-hsl: var(--color-brand-4-hsl);
-  --color-foreground-semi-transparent: hsla(var(--color-brand-4) / 85%);
-  --color-text-on-foreground: var(--color-brand-2);
-  --color-primary: var(--color-brand-1);
-  --color-primary-hsl: var(--color-brand-1-hsl);
-  --color-text-on-primary: var(--color-brand-3);
-  --color-secondary: var(--color-brand-5);
-  --color-secondary-hsl: var(--color-brand-5-hsl);
-  --color-tertiary: var(--color-brand-2);
-  --color-tertiary-hsl: var(--color-brand-2-hsl);
-  --color-text-on-tertiary: var(--color-brand-3);
-  --color-text: var(--color-brand-2);
-  --color-text-hsl: var(--color-brand-2-hsl);
-  --color-text-on-secondary: var(--color-brand-3);
-  --color-text-on-secondary-hsl: var(--color-brand-3-hsl);
-  --color-text-shadow: var(--color-brand-2);
-  --color-border-primary: var(--color-brand-2);
-  --color-border-primary-hsl: var(--color-brand-2-hsl);
-  --color-border-secondary: var(--color-brand-4);
-  --color-border-secondary-hsl: var(--color-brand-4-hsl);
-  --color-border-brand: var(--color-brand);
-  --color-border-brand-hsl: var(--color-brand-hsl);
-  --color-focus-ring: var(--color-brand-5);
-  --color-focus-ring-hsl: var(--color-brand-5-hsl);
-  --color-link: var(--color-brand-1);
-  --focus-ring: 0 0 0 4px var(--color-focus-ring);
-  --color-shadow-hsl: 0 0% 0%;
-  --color-shadow-default: hsl(var(--color-shadow-hsl) / 40%);
-  
-  /* stylelint-disable value-keyword-case */
-  --font-display: 'Caprasimo', 'Caprasimo Fallback', Georgia, serif;
-  --font-copy: system-ui, sans-serif;
-  --font-xl: 2.5rem;
-  --font-lg: 1.9882rem;
-  --font-md: 1.5811rem;
-  --font-sm: 1.2574rem;
-  --font-size: clamp(1rem, .4167vw + .8875rem, 1.125rem);
-  --font-xs: 0.7953rem;
-
-  /* section layout */
-  --section-layout-gap: clamp(var(--spacing-2), 2vw, var(--spacing-4));
-  --section-layout-full: minmax(var(--section-layout-gap), 1fr);
-  --section-layout-content: min((var(--layout-max-width) - var(--section-layout-gap) * 5) / 4, 25% - var(--section-layout-gap));
-  --section-template-columns: 
-    [full-width-start breakout-left-start] var(--section-layout-full)
-    [content-start breakout-right-start content-left-start sidebar-left-start] var(--section-layout-content) [sidebar-left-end content-middle-start] var(--section-layout-content) [content-left-end content-right-start] var(--section-layout-content) [content-middle-end sidebar-right-start] var(--section-layout-content) [content-right-end content-end breakout-left-end]
-    var(--section-layout-full) [breakout-right-end full-width-end];
-}
-
-html.dark-mode {
-  --color-background: var(--color-brand-2);
-  --color-background-hsl: var(--color-brand-2-hsl);
-  --color-foreground: var(--color-brand-4);
-  --color-foreground-semi-transparent: var(--color-brand-4-85);
-  --color-text: var(--color-brand-3);
-  --color-text-on-secondary: var(--color-brand-3);
-  --color-border-primary: var(--color-brand-1);
-  --color-border-secondary: var(--color-brand-4);
-  --color-border-brand: var(--color-brand);
-  --color-focus-ring: var(--color-brand-5);
-  --color-link: var(--color-brand-4);
-}
